+++ java/java.hints/src/org/netbeans/modules/java/hints/errors/Bundle.properties
@@ -197,4 +197,4 @@ FIX_AccessError_PROTECTED=Make {0} protected
 FIX_AccessError_PACKAGE_PRIVATE=Make {0} package private
 ImportClassCustomizer.organizeImports.text=Format and sort imports
 FIX_VarCompDeclaration=Split compound declaration
-FIX_SWITCH_MIX_CASE=Fix mixed case kinds used in switch
+FIX_DifferentCaseKinds=Fix mixed case kinds used in switch
