+++ java/java.hints/src/org/netbeans/modules/java/hints/resources/layer.xml
@@ -326,7 +326,7 @@
 
                 <folder name="bugs">
                     <attr name="SystemFileSystem.localizingBundle" stringvalue="org.netbeans.modules.java.hints.resources.Bundle"/>
-                    <file name="org-netbeans-modules-java-hints-SuspiciousNamesCombination.instance"/>
+                    <file name="org-netbeans-modules-java-hints-SuspiciousNamesCombination.instance" />
                 </folder>
 
                 <folder name="thread">
