+++ java/java.hints/test/unit/src/org/netbeans/modules/java/hints/errors/Bundle_test.properties
@@ -52,6 +52,7 @@ FIX_AccessError_PUBLIC=FIX_AccessError_PUBLIC:{0}
 FIX_AccessError_PROTECTED=FIX_AccessError_PROTECTED:{0}
 FIX_AccessError_PACKAGE_PRIVATE=FIX_AccessError_PACKAGE_PRIVATE:{0}
 FIX_VarCompDeclaration=FIX_VarCompDeclaration
+FIX_DifferentCaseKinds=FIX_DifferentCaseKinds
 
 LBL_Impl_Abstract_Methods=LBL_Impl_Abstract_Methods
 ERR_CannotOverrideAbstractMethods=ERR_CannotOverrideAbstractMethods
